// backend/src/app/setting/setting.interface.ts
export interface ISettingInterface {
  _id?: any;
  currency_symbol?: string;
  currency_code?: string;
  inside_dhaka_shipping_charge?: number;
  outside_dhaka_shipping_charge?: number;
  inside_dhaka_shipping_days?: number;
  outside_dhaka_shipping_days?: number;
  logo?: string;
  favicon?: string;
  title?: string;
  contact?: string;
  email?: string;
  address?: string;
  welcome_message?: string;
  facebook?: string;
  instagram?: string;
  twitter?: string;
  you_tube?: string;
  watsapp?: string;
  about_us?: string;
  return_policy?: string;
  privacy_policy?: string;
  refund_policy?: string;
  cancellation_policy?: string;
  terms_condition?: string;
  shipping_info?: string;
  card_one_logo?: string;
  card_one_title?: string;
  card_two_logo?: string;
  card_two_title?: string;
  card_three_logo?: string;
  card_three_title?: string;
  card_four_logo?: string;
  card_four_title?: string;
}